services/ingestion/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
pydantic==2.12.5
redis==7.1.0
kafka-python==2.0.2
python-dotenv==1.2.1
